# Slightly speed up ActiveJob enqueueing with hashes/keyword arguments