### Motivation / Background

The splat in the case/when is not great since it allocates an array each time it is encountered.

This avoids the allocation, uses a set for the reserved keys, and avoids string allocations for symbol keys.

<details><summary>Benchmark</summary>
<p>

```rb
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "rails", path: "../rails"
  gem "benchmark-ips"
end

require "active_job/railtie"

class TestApp < Rails::Application
  config.load_defaults Rails::VERSION::STRING.to_f
  config.eager_load = false
  config.secret_key_base = "secret_key_base"
  config.active_job.queue_adapter = :test

  config.logger = Logger.new(File::NULL)
end
Rails.application.initialize!

class MyJob < ActiveJob::Base
  def perform
    puts "performed"
  end
end

job_data = ("a".."z").to_h { [it, it] }
Benchmark.ips do |x|
  x.report("perform_later") do
    MyJob.perform_later(job_data)
  end
end
```

</p>
</details> 

Before:
```
ruby 3.4.3 (2025-04-14 revision d0b7e5b6a0) +PRISM [x86_64-linux]
Warming up --------------------------------------
       perform_later     1.541k i/100ms
Calculating -------------------------------------
       perform_later     16.415k (±18.0%) i/s   (60.92 μs/i) -     78.591k in   5.031642s
```

After:
```
ruby 3.4.3 (2025-04-14 revision d0b7e5b6a0) +PRISM [x86_64-linux]
Warming up --------------------------------------
       perform_later     1.834k i/100ms
Calculating -------------------------------------
       perform_later     19.456k (±20.3%) i/s   (51.40 μs/i) -     91.700k in   5.028421s
```

About 18% faster when the hash has many string keys. 

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [ ] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.