@byroot 2025-08-24
Also there's a bit of an issue with using the test adapter, as it keep appending jobs in an array and that stresses the GC a lot.