@Earlopain 2025-08-24
I originally was just benchmarking serialization but thought that was a bit too low-level. If I do just that though, the numbers are even better (almost 50% for the benchmark above, just calling `serialize` on an already instantiated job). Didn't think about how the test adapter works tbh