@byroot 2025-08-24
Updated bechmark:

```ruby
# frozen_string_literal: true
require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "rails", path: "."
  gem "benchmark-ips"
end

require "active_job/railtie"

STAGE = ENV["STAGE"]

module ActiveJob
  module QueueAdapters
    class BlackholeAdapter < AbstractAdapter
      def enqueue(job)
        job.serialize
      end
    end
  end
end

class TestApp < Rails::Application
  config.load_defaults Rails::VERSION::STRING.to_f
  config.eager_load = false
  config.secret_key_base = "secret_key_base"
  config.active_job.queue_adapter = :blackhole

  config.logger = Logger.new(File::NULL)
end
Rails.application.initialize!

class MyJob < ActiveJob::Base
  def perform
    puts "performed"
  end
end

job_data = ("a".."z").to_h { [it.to_sym, it] }

Benchmark.ips do |x|
  x.report(STAGE || "perform_later") do
    MyJob.perform_later(job_data)
  end
  x.compare!(order: :baseline)
  if STAGE
    x.save!("/tmp/bench-aj")
  end
end
```

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after     8.226k i/100ms
Calculating -------------------------------------
               after     84.936k (± 1.8%) i/s   (11.77 μs/i) -    427.752k in   5.037809s

Comparison:
              before:    48629.3 i/s
               after:    84936.3 i/s - 1.75x  faster
```