@Earlopain 2025-08-24
Yup, I noticed that as well. I had a pretty stupid idea for this, assuming all keys are either one or the other is much simpler :sweat_smile:. In practise also often the case I'd say with keyword arguments but never so many as in the benchmark. This was more for when you pass some json over where many keys aren't that uncommon.