@byroot 2025-08-24
> but never so many as in the benchmark.

Yeah, the current benchmark may not be super representative. Still, saving having a copy of all the keys most of the time is a good idea.

If you wish to try that I can hold off.