@Earlopain 2025-08-24
I have a branch here: https://github.com/Earlopain/rails/tree/aj-enqueue-all-symbol-keys

But the numbers aren't really what I would except. Barely any difference or worse, not sure if that is load difference from my system though. It was very hard to measure. The serialization is smaller, but I'd except some improvement in runtime as well (didn't test deserializion)

Feel free to check it out and tweak or just start fresh, your call. I changed the format and content of some of the keys but it should all be backwards compatible. I was testing with this, which is a more realistic benchmark:

<details><summary>Benchmark</summary>
<p>

```rb
# frozen_string_literal: true
require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "rails", path: "../rails"
  gem "benchmark-ips"
end

require "active_job/railtie"

module ActiveJob
  module QueueAdapters
    class BlackholeAdapter < AbstractAdapter
      def enqueue(job)
        job.serialize
      end
    end
  end
end

class TestApp < Rails::Application
  config.load_defaults 8.1
  config.eager_load = false
  config.secret_key_base = "secret_key_base"
  config.active_job.queue_adapter = :blackhole

  config.logger = Logger.new(File::NULL)
end
Rails.application.initialize!

class MyJob < ActiveJob::Base
  def perform
    puts "performed"
  end
end

keys = ["first_job_argument", "second_job_argument", "and_one_more"]
symbols = keys.to_h { [it.to_sym, it] }
strings = keys.to_h { [it, it] }
mixed = strings.transform_keys.with_index { |key, i| i % 2 == 0 ? key : key.to_sym }


prefix = "new-"

Benchmark.ips do |x|
  x.report("#{prefix}symbols") do
    MyJob.perform_later(symbols)
  end

  x.report("#{prefix}strings") do
    MyJob.perform_later(strings)
  end

  x.report("#{prefix}mixed") do
    MyJob.perform_later(mixed)
  end
  x.save!("/tmp/bench-aj")

  x.compare!
end
```

</p>
</details>