@byroot 2025-08-24
What I had in mind is something like this:

```ruby
Benchmark.ips do |x|
  x.report(STAGE || "perform_later") do
    MyJob.perform_later(foo: 1, bar: 2, baz: { egg: 1, spam: 3})
  end
  x.compare!(order: :baseline)
  if STAGE
    x.save!("/tmp/bench-aj")
  end
end
```

That list of argument seem somewhat typical in my opinion.

With a not too complicated diff https://github.com/rails/rails/compare/main...byroot:rails:aj-hash-args (just on the serialization side)

I get:

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after    13.025k i/100ms
Calculating -------------------------------------
               after    133.386k (± 1.0%) i/s    (7.50 μs/i) -    677.300k in   5.078282s

Comparison:
              before:   121386.3 i/s
               after:   133386.2 i/s - 1.10x  faster
```

Which isn't a huge win, but might be worth it?